# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API La Rifa Peru",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://192.168.6.54:8000/api",
            "description": "La Rifa Perú API",
            "variables": {
                "schema": {
                    "enum": [
                        "https",
                        "http"
                    ],
                    "default": "http"
                }
            }
        },
        {
            "url": "http://127.0.0.1:8080/api",
            "description": "localhost8080",
            "variables": {
                "schema": {
                    "enum": [
                        "http"
                    ],
                    "default": "http"
                }
            }
        },
        {
            "url": "http://127.0.0.1/api",
            "description": "localhost",
            "variables": {
                "schema": {
                    "enum": [
                        "http"
                    ],
                    "default": "http"
                }
            }
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Autentificación"
                ],
                "summary": "Autentificación de usuarios",
                "operationId": "1ff9c65c5cd1a8e0786e318eed856972",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "email@gmail.com",
                                    "password": "987654321"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false
            }
        },
        "/auth/storeToken": {
            "post": {
                "tags": [
                    "Autentificación"
                ],
                "summary": "Guardar token de dispositivo",
                "operationId": "f7f98fb110b8166e3bdf4e1ff04f5621",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "token": "admin@gmail.com"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Autentificación"
                ],
                "summary": "Registro de usuario",
                "operationId": "c47a1b563e5e8bc3c7b962f893981e84",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "photo": {
                                        "type": "file"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "email@gmail.com",
                                    "password": "123456789",
                                    "name": "Pepe",
                                    "phone": "phone"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false
            }
        },
        "/auth/info": {
            "get": {
                "tags": [
                    "Autentificación"
                ],
                "summary": "Información del usuario logeado",
                "operationId": "eb02f0c18032f15a98791041bbb6d9c3",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Autentificación"
                ],
                "summary": "Finaliza la sessión del usuario",
                "operationId": "eae6a8251a6469bbabe454860f64cf37",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/users/{id_user}": {
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Modifiy user information",
                "operationId": "1afc67ddeed0ff529a72d7f249b5c1d8",
                "parameters": [
                    {
                        "name": "id_user",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "photo": {
                                        "type": "file"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Pepe",
                                    "surnames": "Garcia Fuentes",
                                    "phone": "984575821"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/users/public/resetpass/{email}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "email to reset password, send code to email",
                "operationId": "93b785331fd73faddb226c2794d8cb68",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false
            }
        },
        "/users/public/resetpass": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Reset password",
                "operationId": "9778d67f89a9fe5ae52d293d6f75727c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "password": "123456",
                                    "token": "123456"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false
            }
        },
        "/home": {
            "get": {
                "tags": [
                    "Home"
                ],
                "summary": "Retorna los datos a mostrar en el inicio",
                "operationId": "f8dc767a4c83166e4cdcbccd11d666f8",
                "responses": {
                    "200": {
                        "description": "Home page data"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/category/{id_category}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Retorna las rifas por categoria",
                "operationId": "a36233fb635ebb40bff4f7102027dc35",
                "parameters": [
                    {
                        "name": "id_category",
                        "in": "path",
                        "description": "id de categoria",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Page category data"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/winners": {
            "get": {
                "tags": [
                    "Home"
                ],
                "summary": "Retorna los datos a mostrar en el inicio",
                "operationId": "fd8638b9c318a4adae18161f8d2db44c",
                "responses": {
                    "200": {
                        "description": "Home page data"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/detailRaffle/{id_raffle}": {
            "get": {
                "tags": [
                    "Home"
                ],
                "summary": "Retorna el detalle de una rifa.",
                "operationId": "b1ca179ce36b6805be958ab48a4d7b36",
                "parameters": [
                    {
                        "name": "id_raffle",
                        "in": "path",
                        "description": "id de rifa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Home page data"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/favoritesRaffles": {
            "get": {
                "tags": [
                    "Home"
                ],
                "summary": "Retorna tus rifas favoritas.",
                "operationId": "2cfc09192af7a2d33675e1ccf375dc15",
                "responses": {
                    "200": {
                        "description": "Home page data"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/users/favorites": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "New favorite",
                "operationId": "9f4efa7ba1b4295adc6092cfd48e4c42",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_raffle": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id_raffle": "123456"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/users/favorites/{id_raffle}": {
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Delete favorite",
                "operationId": "9e4443bdba9ab09a8718ed699105c36e",
                "parameters": [
                    {
                        "name": "id_raffle",
                        "in": "path",
                        "description": "id de rifa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "login",
                "name": "token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}